Resolve #1: move redis thing in domain pkg to other file
